board_env/board_project/requirements.txt
Django==5.2.4
PyMySQL==1.1.0
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.5.0
gunicorn==21.2.0